add notification when user update card
also allows user to edit page is they are the one who made pending edits
